app/requirements.txt
Django==4.1.7
psycopg2==2.8.6
djangorestframework==3.14.0